add ColorscaleSelect and ScaleSelect components for controls
