+++ src/lib/components/dotplot/DotplotControls.js
@@ -1,14 +1,12 @@
 import React, { useEffect, useState } from "react";
 
-import _ from "lodash";
 import { Button, Form, InputGroup } from "react-bootstrap";
 
-import { COLORSCALES } from "../../constants/colorscales";
-import { DOTPLOT_SCALES } from "../../constants/constants";
 import {
   useSettings,
   useSettingsDispatch,
 } from "../../context/SettingsContext";
+import { ColorscaleSelect, ScaleSelect } from "../controls/Controls";
 
 export function DotplotControls() {
   const settings = useSettings();
@@ -39,24 +37,7 @@ export function DotplotControls() {
   return (
     <>
       <Form>
-        <Form.Group className="mb-1">
-          <Form.Label>Colorscale</Form.Label>
-          <Form.Select
-            value={settings.controls.colorScale}
-            onChange={(e) => {
-              dispatch({
-                type: "set.controls.colorScale",
-                colorScale: e.target.value,
-              });
-            }}
-          >
-            {_.keys(COLORSCALES).map((key) => (
-              <option key={key} value={key}>
-                {key}
-              </option>
-            ))}
-          </Form.Select>
-        </Form.Group>
+        <ColorscaleSelect />
         <Form.Group className="mb-2">
           <InputGroup>
             <InputGroup.Text>min</InputGroup.Text>
@@ -122,25 +103,7 @@ export function DotplotControls() {
             </Button>
           </InputGroup>
         </Form.Group>
-        <Form.Group className="mb-2">
-          <Form.Label>Standard scale</Form.Label>
-          <Form.Select
-            value={settings.controls.scale.dotplot || ""}
-            onChange={(e) => {
-              dispatch({
-                type: "set.controls.scale",
-                plot: "dotplot",
-                scale: !e.target.value.length ? null : e.target.value,
-              });
-            }}
-          >
-            {_.values(DOTPLOT_SCALES).map((scale) => (
-              <option key={scale.value} value={scale.value || ""}>
-                {scale.name}
-              </option>
-            ))}
-          </Form.Select>
-        </Form.Group>
+        <ScaleSelect plot="dotplot" />
         <Form.Group className="mb-2">
           <Form.Label>Expression Cutoff</Form.Label>
           <InputGroup>
